\name{CpmLogLogisticPMITICS-package}
\alias{CpmLogLogisticPMITICS-package}
\alias{CpmLogLogisticPMITICS}
\docType{package}
\title{
\packageTitle{CpmLogLogisticPMITICS}
}
\description{
\packageDescription{CpmLogLogisticPMITICS}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{CpmLogLogisticPMITICS}

Maintainer: \packageMaintainer{CpmLogLogisticPMITICS}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = CpmLogLogisticPMITICS)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{CpmLogLogisticPMITICS}}
%\section{Documentation Index}{\packageIndices{CpmLogLogisticPMITICS}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
